requirements.txt
ansi2html==1.8.0
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dash==2.11.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et-xmlfile==1.1.0
Flask==2.2.5
frozendict==2.3.8
gunicorn==21.2.0
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
multitasking==0.0.11
nest-asyncio==1.5.6
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
plotly==5.15.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
retrying==1.3.4
sequential==1.0.0
six==1.16.0
soupsieve==2.4.1
tenacity==8.2.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
webencodings==0.5.1
Werkzeug==2.2.3
yfinance==0.2.27
